# flag-backend-master/src/main/java/ph/devcon/flag/entrypoint/api/rest/sms/SendSmsMessageRequest.java
package ph.devcon.flag.entrypoint.api.rest.sms;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;


@Data
@NoArgsConstructor
@AllArgsConstructor
public final class SendSmsMessageRequest {

    private String number;
    private String message;

}