# flag-backend-master/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  title: DCTx FLAG API
  description: >-
    API for FLAG Web Services
  version: "1.0.0"
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - BearerAuth: []
paths:
...
  /sms/message:
    post:
      tags:
        - requestor
        - dashboard
      summary: >-
        Endpoint to send an SMS message. This requires a mobile number and a message. This
        will return the message reference ID and the status.
      operationId: sendSmsMessage
      requestBody:
        description: >-
          SMS message request body contains the mobile number and a short message (160 ASCII characters)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendSmsMessageRequest'
      responses:
        200:
          description: >-
            A successful request will return the message status (Pending, Sent, or Failed) and the message reference ID,
            which can used to check if the status has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendSmsMessageResponse'
        default:
          description: >-
            An error occured while performing the operation. A detailed JSON error will be returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sms/message/{id}/status:
    get:
      tags:
        - requestor
        - dashboard
      summary: >-
        Retrieve the status of a message
      operationId: getMessageStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: the ID of the message
            example: '80404765'
      responses:
        200:
          description: >-
            The status of the message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendSmsMessageResponse'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /donors:
    get:
      tags:
        - donation
      summary: >-
        Retrieve the details of a donor
      operationId: getDonors
      parameters:
      responses:
        200:
          description: >-
            The list of all donors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donor'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - donation
      summary: Creates a donor.
      operationId: createDonor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Donor'
      responses:
        200:
          description: >-
            Donor created
          content:
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /donors/{id}:
    get:
      tags:
        - donation
      summary: >-
        Retrieve the details of a donor by id
      operationId: getDonorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: the ID of the donor
            example: '1'
      responses:
        200:
          description: >-
            The details of the donor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donor'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /donors/mobile/{number}:
    get:
      tags:
        - donation
      summary: >-
        Retrieve the details of a donor by mobile number
      operationId: getDonorByMobile
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: string
            description: the mobile number of the donor
            example: '630000000000'
      responses:
        200:
          description: >-
            The details of the donor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donor'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /donors/sector/{id}:
    get:
      tags:
        - donation
      summary: >-
        Retrieve the details of a donor by sector
      operationId: getDonorBySector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            description: Sector classification of the donor
            example: 1
      responses:
        200:
          description: >-
            The details of the donor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donor'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /donationrequests/donor/{id}:
    get:
      tags:
        - donation
      summary: >-
        Retrieve all donation request of a donor
      operationId: getDonationRequestByDonorid
      parameters:
          - name: id
            in: path
            required: true
            type: number
            description: Id of the donor
            example: 1
      responses:
        200:
          description: >-
            The details of the requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationRequest'
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /reports:
    get:
      tags:
        - donation
      summary: >-
        Generate donation reports
      operationId: generateReports
      responses:
        200:
          description: >-
            Generated successfully
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reports/{year}:
    get:
      tags:
        - donation
      summary: >-
        Generate donation reports by year
      operationId: generateReportsByYear
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: number
            description: the year
            example: 2020
      responses:
        200:
          description: >-
            Generated successfully
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        id:
          type: string
          description: >-
            a unique identifier for this particular occurrence of the problem.
        links:
          type: object
          description: >-
            a links object containing the following members;
          properties:
            about:
              type: string
              description: >-
                a link that leads to further details about this particular occurrence of the problem.
          additionalProperties: true
        status:
          type: string
          description: >-
            the HTTP status code applicable to this problem, expressed as a
            string value.
        code:
          type: string
          description: >-
            an application-specific error code, expressed as a string value.
        title:
          type: string
          description: >-
            a short, human-readable summary of the problem that SHOULD NOT
            change from occurrence to occurrence of the problem, except for
            purposes of localization.
        detail:
          type: string
          description: >-
            a human-readable explanation specific to this occurrence of the
            problem. Like title, this field’s value can be localized.
        source:
          type: array
          description: >-
            an object containing references to the source of the error, optionally including any of the following members;
          items:
            type: object
            properties:
              pointer:
                type: string
                description: >-
                  a JSON Pointer [RFC6901] to the associated entity in the
                  request document [e.g. "/data" for a primary data object, or "
                  /data/attributes/title" for a specific attribute].
              parameter:
                type: string
                description: >-
                  a string indicating which URI query parameter caused the error
                  .
        meta:
          type: object
          description: >-
            a meta object containing non-standard meta-information about the error.
          additionalProperties: true
...
    SendSmsMessageRequest:
      type: object
      properties:
        number:
          type: string
          description: >-
            The mobile number to send the SMS message.
          example: '+630000000000'
        message:
          type: string
          description: >-
            The short message (max 160 ASCII characters) to send.
          example: 'This is the message.'
    SendSmsMessageResponse:
      type: object
      properties:
        message_id:
          type: string
          description: >-
            The reference ID for the message that can be used to check its updated status later.
        status:
          type: string
          description: >-
            The status of the message (Pending, Sent, or Failed)
    Donor:
      type: object
      properties:
        id:
          type: integer
          description: >-
            The ID of the donor
          example: 1
        contact_person:
          type: string
          description: >-
            The donor's name or representative of an affiliation or organization
        affiliation_org:
          type: string
          description: >-
            The name organization or affiliation giving the donation
        mobile_number:
          type: string
          description: >-
            The donor's mobile number
          example: '+639990001111'
        landline_number:
          type: string
          description: >-
            The donor's landline number
          example: '+639990001111'
        isAnonymous:
          type: boolean
          description: >-
            Indicate if donor wanted to be anonymous
          example: 'false'
        isInternational:
          type: boolean
          description: >-
            Indicate if donor is from outside the country
          example: 'false'
        sector_type:
          type: string
          description: >-
            The donor type
          enum:
            - GOV
            - LGU
            - PRINGO
            - INGO
          example: GOV
        email_address:
          type: string
          description: >-
            The donor's email address
          format: email
          example: donor@acme.corp
        address:
          type: object
          description: >-
            This contains the street no., block, house number etc.
          example: {
                    "address_line_one": "Block1, AFPOVAI",
                    "barangay": "Western Bicutan",
                    "city": "Taguig",
                    "province": "Metro Manila",
                    "region": "Region VII",
                    "country": "Philippines"
                  }
...